Use correct cert_store for creating the database certificate chain
We already have our own cert_store in the engine as exclusive store for
databases loaded from a file, and the NULL engine used otherwise uses
the same certificates as our cert_store uses (hopefully).

We need to pass the cert_store with all the issuer certificates in the
certificate chain passed to the verify function, otherwise building the
chain is going to fail.